Migration from SWAT-Research
Brings 793 commits of development work including:

- Symbolic execution enhancements and new dispatcher architecture
- Spring framework instrumentation support
- Symbolic array handling and annotation-based symbolization
- Enhanced constraint solving with caching
- New analysis modules and coverage tracking
- HTTP driver and SV-COMP integration
- Z3 solver library updates (4.12.5 -> 4.15.4)
- Additional target applications and test cases

+++ symbolic-executor/src/main/java/de/uzl/its/swat/symbolic/trace/SpecialElement.java
@@ -19,7 +19,7 @@ class SpecialElement extends Element {
      * @param iid The unique identifier of the element
      * @param inst The string representation of the instruction that caused the special element
      */
-    SpecialElement(int iid, String inst) {
+    SpecialElement(long iid, String inst) {
         this.setIid(iid);
         this.inst = inst;
     }
